requirements.txt
qwen-tts
stable-ts
faster-whisper
soundfile
numpy
torch
pillow
